+++ components/nfc/tools/api/src/main/java/com/flipperdevices/nfc/tools/api/NfcToolsApi.kt
@@ -1,5 +1,7 @@
 package com.flipperdevices.nfc.tools.api
 
+import java.math.BigInteger
+
 interface NfcToolsApi {
-    suspend fun bruteforceKey(mfKey32Nonce: MfKey32Nonce): ULong?
+    suspend fun bruteforceKey(mfKey32Nonce: MfKey32Nonce): BigInteger?
 }
